pyproject.toml
[tool.poetry]
name = "aidoctool"
version = "0.1.0"
description = ""
authors = ["JV <viloforge@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
pyyaml = "*"
python-dotenv = "*"
click = "*"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
